[refactor] InvestmentFactory 리팩토링 (#29)
* rename interface name - Investment to ExpirationInvestment

객체가 계산되는 정보가 만기(Expiration) 기준으로 계산되기 때문에 이름을 변경함

* rename class name - DefaultInvestmentFactory to ExpirationInvestmentFactory

* remove static keyword

* formatting code - change method order

* rename method name - getAccumulatedPrincipal to getPrincipal

* rename method name - getAccumulatedInterest to getInterest

* rename method name - getAccumulatedTax to getTax

* rename method name - getAccumulatedTotalProfit to getTotalProfit

* change MonthlyInvestmentResult type - class to record

* formatting code

* add interface and imple class - InvestmentCalculator, ExpirationInvestmentCalculator

* change method signiture - ExpirationInvestment

해당 타입은 임시로 추가한 것으로 차후 Investment로 통합될 예정

* add test - calPrincipal

* introduce field - InvestmentCalculator

* add test - CompoundFixedDeposit case

* rename interface name - ExpirationInvestment to Investment

변경한 이유는 만기와 월회차 기간에 대한 기준을 Calculator 객체에게 위임하기 위해서이다.

* add test - calPrincipal, CompoundFixedInstallmentSaving instance

* remove Calculator class and interface

* rename method name - getPrincipalAmount to getPrincipal

* add method in Investment - getPrincipal(int)

* rename method name - getAmount to getTotalProfit

* fix getInterest(int) method logic

* add getTax(int) method

* formatting Investment code

* extract method - applyTax(int)

* formatting method order

* introduce variable - getTotalProfit

* add comment - Investment

* delete implements MonthlyInvestment in SimpleFixedDeposit

* add test - getPrincipal

* add test - getPrincipal(int)

* add test - getInterest

* add test - calMonthlyInterest

* change method signiture - int to double

* update calMonthlyInterest method to return double; adjust related tests and output values

* change production code's return type - double to BigDecimal

* change return type - double to BigDecimal

* update calMonthlyInterest methods to return double; adjust calculations to use BigDecimal

* add commtnet and change return type - getMonthlyRate, double to BigDecimal

* update code - using BigDecimal

- BigDecimal 객체를 사용하여 반환하도록 함
- scale=6으로 설정하여 소수점 아래 6자리까지 표현하고, 7번째 자리는 반올림하도록 정책을 설정함.
- 추후 scale 변수 값은 리팩토링하여 어느 한군데에 배치할 예정

* update code - using doubleValue method

- 반환된 타입이 BigDecimal 객체를 doubleValue() 메서드를 사용하여 double 값으로 변환한다.
- 추후 다른 메서드 타입도 BigDecimal 타입으로 변경하기 위해서 임시적으로 doubleValue()를 호출함

* change test assert BigDecimal method

- BigDecimal 객체들을 대상으로 대소비교하기 위해서 delta 추가
- delta를 이용하여 소수점 아래 6자리까지 같은지 비교함.
  - ex) a=0.004167, b=0.0041666 은 같은 값으로 취급함

* update README.md

소수점 관련 처리 내용 추가

* change expected value

* add util class - BigDecimalUtils

* change return type - getAnnualRate(), double to BigDecimal

* change method return type - double to BigDecimal

* add test - testValueOf

* add round

* add comment - InterestRate

* change method return type - double to BigDeicmal

* delete delta variable

* add testutil - BigDecimalAssertion

해당 클래스는 테스트 검증을 유용하게 하기 위해서 만든 클래스이다.

* change return type - double to BigDecimal

* fix code - getInterest()

* fix test

* add todo comment

* update code - use DECIMAL64

- DECIMAL64는 소수점 아래 최대 16자리까지 표시하고, 은행가 반올림(HALF_EVEN)을 사용한다.
- 은행가 반올림은 반올림할 자리의 다음 숫자가 0~4이면 버리고, 6~9이면 올린다. 5인 경우에 반올림할 자리의 다음 숫자(5) 뒤에 없거나 0인 경우에 반올림할 자리의 숫자가 짝수이면 내림하고, 홀수이면 올림한다.
- ex) 1.235는 반올림할 자리의 숫자(0.03)가 홀수이므로 올려서 1.24가 된다.
- ex) 1.245는 반올림할 자리의 숫자(0.04)가 짝수이므로 버림하여 1.24가 된다.
- ex) 1.234는 반올림할 자리의 다음 숫자(0.004)가 5보다 작기 때문에 버림한다.
- ex) 1.236는 반올림할 자리의 다음 숫자(0.006)가 5보다 크기 때문에 올림한다.

* change update code - getInterest(int)

- 반복문 방식이 아닌 원금에서 곱하는 방식으로 계산함
- 성장계수를 구하는 calTotalGrowthFactor와 getGrowthFactor가 매개변수만 다르지 헷갈리는 부분이 있어서 메서드를 하나로 합치거나 오버로딩할 예정

* fix test

- 복리 이자 계산 방식이 변경됨에 따라서 관련된 테스트의 예상값도 변경됨

* delete BigDecimalUtil class

* fix code - getInterest(int)

정수형으로 반환시 뱅커 라운딩 방식으로 반올림함

* refactor method - calTotalGrowthFactor

+++ src/main/java/domain/amount/FixedDepositAmount.java
@@ -1,5 +1,7 @@
 package domain.amount;
 
+import java.math.BigDecimal;
+
 import domain.interest_rate.InterestRate;
 
 public class FixedDepositAmount implements LumpSumInvestmentAmount {
@@ -25,11 +27,11 @@ public class FixedDepositAmount implements LumpSumInvestmentAmount {
 
 	@Override
 	public double calAnnualInterest(InterestRate interestRate) {
-		return interestRate.getAnnualInterest(amount);
+		return interestRate.getAnnualInterest(amount).doubleValue();
 	}
 
 	@Override
-	public int calMonthlyInterest(InterestRate interestRate) {
+	public BigDecimal calMonthlyInterest(InterestRate interestRate) {
 		return interestRate.calMonthlyInterest(amount);
 	}
 }
