+++ src/main/java/domain/amount/MonthlyInstallmentInvestmentAmount.java
@@ -1,5 +1,7 @@
 package domain.amount;
 
+import java.math.BigDecimal;
+
 import domain.interest_rate.InterestRate;
 
 public class MonthlyInstallmentInvestmentAmount implements InstallmentInvestmentAmount {
@@ -20,11 +22,11 @@ public class MonthlyInstallmentInvestmentAmount implements InstallmentInvestment
 
 	@Override
 	public double calAnnualInterest(InterestRate interestRate) {
-		return interestRate.getAnnualInterest(amount);
+		return interestRate.getAnnualInterest(amount).doubleValue();
 	}
 
 	@Override
-	public int calMonthlyInterest(InterestRate interestRate) {
+	public BigDecimal calMonthlyInterest(InterestRate interestRate) {
 		return interestRate.calMonthlyInterest(amount);
 	}
 }
